refactor(src): split daemon/orchestration.rs by responsibility
